commented out food truck add event code. Submit event working
