+++ package.json
@@ -11,6 +11,7 @@
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.9.0",
     "@mui/x-date-pickers": "^5.0.0-beta.1",
+    "@reach/combobox": "^0.17.0",
     "@react-google-maps/api": "^2.12.0",
     "axios": "^0.27.2",
     "dotenv": "^16.0.1",
