+++ src/components/auth/SignupVenue.jsx
@@ -18,6 +18,7 @@ import {signUpVenue} from 'services/authServices';
 import IconButton from '@mui/material/IconButton';
 import PhotoCamera from '@mui/icons-material/PhotoCamera';
 import PlacesAutocomplete, {geocodeByAddress, getLatLng} from 'react-places-autocomplete';
+import AutoComplete from 'components/map/AutoComplete';
 
 function Copyright(props) {
   return (
@@ -47,31 +48,27 @@ function SignupVenue() {
     google_maps: '',
     mobile: '',
     password: '',
-    password_confirmation: ''
-    // address_attributes: {
-    //   street: '',
-    //   suburb: '',
-    //   state: 'Queensland',
-    //   postcode: '',
-    //   country: 'Australia'
-    // }
-  };
-  // used by get coordinates from google maps
-  const [address, setAddress] = useState();
-  const [coordinates, setCoordinates] = useState({
+    password_confirmation: '',
+    address: '',
     lat: null,
     lng: null
-  });
+  };
+  // used by get coordinates from google maps
+  // const [address, setAddress] = useState();
+  // const [coordinates, setCoordinates] = useState({
+  //   lat: null,
+  //   lng: null
+  // });
 
   const [formData, setFormData] = useState(initialFormData);
   const [error, setError] = useState(null);
 
-  const handleSelect = async (value) => {
-    const results = await geocodeByAddress(value);
-    const latLng = await getLatLng(results[0]);
-    setAddress(value);
-    setCoordinates(latLng);
-  };
+  // const handleSelect = async (value) => {
+  //   const results = await geocodeByAddress(value);
+  //   const latLng = await getLatLng(results[0]);
+  //   setAddress(value);
+  //   setCoordinates(latLng);
+  // };
 
   console.log(formData);
 
@@ -184,65 +181,8 @@ function SignupVenue() {
                 />
               </Grid>
               This is the address form
-              {/* <Grid item xs={12}>
-                <TextField
-                  required
-                  id="street"
-                  name="street"
-                  label="Address"
-                  fullWidth
-                  autoComplete="address-line-1"
-                  value={formData.address_attributes.value}
-                  onChange={handleAddressFormData}
-                />
-                <TextField
-                  required
-                  id="suburb"
-                  name="suburb"
-                  label="Suburb"
-                  fullWidth
-                  autoComplete="suburb"
-                  value={formData.address_attributes.value}
-                  onChange={handleAddressFormData}
-                />
-                <TextField
-                  required
-                  id="postcode"
-                  name="postcode"
-                  label="Postcode"
-                  fullWidth
-                  autoComplete="postcode"
-                  value={formData.address_attributes.value}
-                  onChange={handleAddressFormData}
-                />
-              </Grid> */}
               <Grid item xs={12}>
-                <PlacesAutocomplete value={address} onChange={setAddress} onSelect={handleSelect}>
-                  {({getInputProps, suggestions, getSuggestionItemProps, loading}) => (
-                    <div>
-                      <p>Latitude: {coordinates.lat}</p>
-                      <p>Longitude: {coordinates.lng}</p>
-
-                      <input {...getInputProps({placeholder: 'Type address'})} />
-
-                      <div>
-                        {loading ? <div>...loading</div> : null}
-
-                        {suggestions.map((suggestion) => {
-                          const style = {
-                            backgroundColor: suggestion.active ? '#41b6e6' : '#fff'
-                          };
-
-                          return (
-                            <div key="id" {...getSuggestionItemProps(suggestion, {style})}>
-                              {suggestion.description}
-                            </div>
-                          );
-                        })}
-                      </div>
-                    </div>
-                  )}
-                </PlacesAutocomplete>
+                <AutoComplete name="address" required id="address" />
               </Grid>
               <Grid item xs={12} sm={6}>
                 <TextField
